// src/types/ISensorInfo.tsx
// SPDX-FileCopyrightText: 2024 Osnabrück University of Applied Sciences
// SPDX-FileContributor: Andreas Schliebitz
// SPDX-FileContributor: Henri Graf
// SPDX-FileContributor: Jonas Tüpker
// SPDX-FileContributor: Lukas Hesse
// SPDX-FileContributor: Maik Fruhner
// SPDX-FileContributor: Prof. Dr.-Ing. Heiko Tapken
// SPDX-FileContributor: Tobias Wamhof
//
// SPDX-License-Identifier: MIT

export enum SensorType {
  CAMERA = "camera"
}

export default interface ISensorInfo {
    type: SensorType;
    name: string;
    manufacturer: string;
    model: string;
    serial: string;
    hostname: string;
    ip: string;
}